feat(recipe): 量/単位入力を単位タブ+数量ドロップダウンに刷新しモバイル UI を調整 (#36)
## 量/単位の入力刷新
- 自由入力(数値+単位テキスト)をやめ、単位をタブ(チップ)で選択 → 単位に応じた
  数量候補のドロップダウンから選ぶ方式に。単位ごとに開始値・刻み・範囲を変える
  (g は 100 から 10 刻み、大さじは 1 から 0.5 刻み、個は 1 刻み、適量・少々は数量なし)。
- 分数対応: quantity を integer→number(小数) に。API(openapi)・Go domain(float64/numeric)・
  DTO・zod を追従。大さじ1/2=0.5 を保存・表示できる。
- 表示: 0.5→½、1.5→1½ に整形。大さじ系は単位を前(大さじ½)、g/個系は後ろ(100g/1個)。
  適量・少々は単位のみ表示。単位設定は frontend/src/features/recipes/units.ts に集約。

## モバイル UI 調整
- ヘッダーの横パディングを px-3 sm:px-4 lg:px-6 にし、端の詰まりを解消(カードと余白を揃える)。
- 新規作成フォームで、スマホ時に人数と調理時間を横並びにする(sm 以上は従来の3列)。

実 DB で 大さじ0.5(numeric)・適量 を確認。バックエンド build/unit/結合、フロント
typecheck/lint/vitest/Playwright すべて緑。

+++ api/internal/apigen/apigen.gen.go
@@ -10,14 +10,14 @@ const (
 // CookingInput defines model for CookingInput.
 type CookingInput struct {
 	Ingredients NameInput `json:"ingredients"`
-	Quantity    int       `json:"quantity"`
+	Quantity    float64   `json:"quantity"`
 	Unit        string    `json:"unit"`
 }
 
 // CookingResponse defines model for CookingResponse.
 type CookingResponse struct {
 	Ingredients NameResponse `json:"ingredients"`
-	Quantity    int          `json:"quantity"`
+	Quantity    float64      `json:"quantity"`
 	Unit        string       `json:"unit"`
 }
 
@@ -100,14 +100,14 @@ type RegisterRequest struct {
 
 // SeasonInput defines model for SeasonInput.
 type SeasonInput struct {
-	Quantity  int       `json:"quantity"`
+	Quantity  float64   `json:"quantity"`
 	Seasoning NameInput `json:"seasoning"`
 	Unit      string    `json:"unit"`
 }
 
 // SeasonResponse defines model for SeasonResponse.
 type SeasonResponse struct {
-	Quantity  int          `json:"quantity"`
+	Quantity  float64      `json:"quantity"`
 	Seasoning NameResponse `json:"seasoning"`
 	Unit      string       `json:"unit"`
 }
